♻️[Refactor] 컨테이너 배럴 파일 삭제
