+++ src/app/bookclub/create/page.tsx
@@ -1,4 +1,4 @@
-import { CreateBookClubPage } from '@/features/club-create/container';
+import CreateBookClubPage from '@/features/club-create/container/CreateBookClubPage';
 
 export default function BookClubCreate() {
   return <CreateBookClubPage />;
